[front] feat: add pptx geometry feebacks
